:truck: Extract logic into reorderCheckboxes.ts
